% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{hook_plot_wired}
\alias{hook_plot_wired}
\title{Wired plot hooks for HTML document}
\usage{
hook_plot_wired(x, options)
}
\arguments{
\item{x}{Filename for the plot (a character string).}

\item{options}{A list of the current chunk options.}
}
\value{
A character string of code, with plot filenames wrapped.
}
\description{
Wired plot hooks for HTML document
}
\examples{
# todo
}
