% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wired-package}
\alias{wired-package}
\title{Sketchy components and charts}
\description{
Hand drawn UI elements with \href{https://wiredjs.com}{wired}
 and sketchy charts with \href{https://timqian.com/chart.xkcd}{chart.xkcd}
}
\author{
Victor Perrier (@dreamRs_fr)
}
