% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{wired-progress}
\alias{wired-progress}
\alias{wired_progress}
\alias{update_wired_progress}
\title{Wired Progress bar}
\usage{
wired_progress(id, value, min = 0, max = 100, percentage = TRUE,
  width = "100\%")

update_wired_progress(id, value,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Id for the progress bar.}

\item{value}{Numeric value of the progress.}

\item{min}{Minimum value. Default is 0.}

\item{max}{Maximum value. Default is 100.}

\item{percentage}{Boolean indicating if the label should show a \% symbol.}

\item{width}{Width of the progress bar.}

\item{session}{Shiny session.}
}
\value{
an HTML tag
}
\description{
Wired Progress bar
}
\examples{

wired_progress("ID", value = 50)

wired_progress("ID", value = 150, max = 200, percentage = FALSE)


if (interactive()) {
  library(shiny)
  library(wired)

  ui <- fluidPage(
    tags$h2("Wired progress bar"),
    wired_progress("progress", value = 10),
    numericInput("update", "Enter new value:", 0)
  )

  server <- function(input, output, session) {

    observe({
      update_wired_progress("progress", input$update)
    })

  }

  shinyApp(ui, server)
}
}
