% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{wired_button}
\alias{wired_button}
\title{Wired Action button}
\usage{
wired_button(inputId, label = NULL, icon = NULL, width = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value. The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label.}

\item{icon}{An optional icon to appear on the button.}

\item{width}{The width of the input, e.g. \code{400px}, or \code{100\%}; see \link[htmltools]{validateCssUnit}.}

\item{...}{Named attributes to be applied to the button.}
}
\description{
Wired Action button
}
\examples{
\dontrun{

if (interactive()) {

library(shiny)
library(wired)

ui <- fluidPage(
  wired_button(
    inputId = "go",
    label = "GO!"
  ),
  tags$br(),
  verbatimTextOutput(outputId = "res")
)

server <- function(input, output, session) {

  output$res <- renderPrint(input$go)

}

shinyApp(ui, server)

}

}
}
