% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{wired_card}
\alias{wired_card}
\title{Wired Card}
\usage{
wired_card(..., elevation = 2, fill = NULL, padding = 10, width = NULL)
}
\arguments{
\item{...}{UI objects to be contained in the card.}

\item{elevation}{Numerical number between 1-5 (inclusive) -
sets the elevation of the card. Default is 2.}

\item{fill}{A color to fill the background of the card in a sketchy format.}

\item{padding}{Amount of padding to apply. Can be numeric
(in pixels) or character (e.g. "3em").}

\item{width}{The width of the input, e.g. \code{400px},
or \code{100\%}; see \link[htmltools]{validateCssUnit}.}
}
\description{
Wired Card
}
\examples{
library(htmltools)
browsable(tagList(
  wired_card(
    "A sketchy card !"
  ),
  tags$br(),
  "Another with fill background:",
  wired_card(
    "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
    " Vestibulum imperdiet, erat a porttitor sagittis,",
    " eros mauris finibus diam, porta venenatis enim leo ut libero.",
    " Morbi volutpat nibh ex, quis faucibus ante dictum vestibulum.",
    " Etiam cursus turpis in lorem rutrum, vitae imperdiet sem mollis.",
    " Mauris quis neque ac elit suscipit posuere.",
    fill = "red", width = "100\%"
  )
))

if (interactive()) {

  library(shiny)
  library(wired)

  ui <- fluidPage(
    fluidRow(
      column(
        width = 4,
        wired_card("Simple card")
      ),
      column(
        width = 4,
        wired_card(
          "Another one with a plot!",
          plotOutput(outputId = "plot"),
          elevation = 5, width = "100\%"
        )
      )
    )
  )

  server <- function(input, output, session) {

    output$plot <- renderPlot({
      plot(sin(seq(-pi, pi, length.out = 100)), type = "l")
    })

  }

  shinyApp(ui, server)
}
}
