% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{wired_divider}
\alias{wired_divider}
\title{Wired Divider}
\usage{
wired_divider(elevation = 1)
}
\arguments{
\item{elevation}{Numerical number between 1-5 (inclusive) -
sets the elevation of the card. Default is 1.}
}
\value{
an HTML tag
}
\description{
Wired Divider
}
\examples{
library(htmltools)
browsable(tagList(
  "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
  wired_divider(),
  "Quisque at nunc nulla. Nulla tincidunt est sed efficitur fermentum.",
  wired_divider(2),
  "In consequat id elit sed tristique.",
  wired_divider(4),
  "Ut vel vulputate quam. Suspendisse non."
))
}
