% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{wired_document}
\alias{wired_document}
\title{Convert to a Wired HTML document}
\usage{
wired_document(...)
}
\arguments{
\item{...}{Params}
}
\value{
R Markdown output format to pass to render.
}
\description{
Convert to a Wired HTML document
}
\examples{
# todo
}
