% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{wired_image}
\alias{wired_image}
\title{Wired Image}
\usage{
wired_image(src, elevation = 1, ...)
}
\arguments{
\item{src}{URL / path of the image.}

\item{elevation}{Numerical number between 1-5 (inclusive) -
sets the elevation of the card. Default is 1.}

\item{...}{Named attributes to be applied to the image.}
}
\value{
an HTML tag
}
\description{
Wired Image
}
\examples{
library(htmltools)
wired_image(
  src = "https://www.gstatic.com/webp/gallery/1.sm.jpg"
)

wired_image(
  src = "https://www.gstatic.com/webp/gallery/1.sm.jpg",
  elevation = 5
)
}
