% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{wired_radio}
\alias{wired_radio}
\title{Wired Radio Buttons}
\usage{
wired_radio(inputId, label, choices, selected = NULL, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from. If elements of the list are named, then that name rather than the value is displayed to the user.}

\item{selected}{The initially selected value. If not specified then defaults to the first value.}

\item{width}{The width of the input, e.g. \code{400px}, or \code{100\%}; see \link[htmltools]{validateCssUnit}.}
}
\description{
Wired Radio Buttons
}
\examples{
if (interactive()) {

  library(shiny)
  library(wired)

  ui <- fluidPage(
    wired_radio(
      inputId = "radi", label = "Radio buttons:",
      choices = c("Normal" = "norm",
                  "Uniform" = "unif",
                  "Log-normal" = "lnorm",
                  "Exponential" = "exp")
    ),
    verbatimTextOutput(outputId = "res")
  )

  server <- function(input, output, session) {

    output$res <- renderPrint(input$radi)

  }

  shinyApp(ui, server)

}
}
