% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{wired_search}
\alias{wired_search}
\title{Wired Search Input}
\usage{
wired_search(inputId, label = NULL, placeholder = NULL, value = "",
  width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control.}

\item{value}{Initial value for the input.}

\item{width}{The width of the input, e.g. \code{400px}, or
\code{100\%}; see \link[htmltools]{validateCssUnit}.}
}
\description{
Wired Search Input
}
\examples{
if (interactive()) {

  library(shiny)
  library(wired)

  ui <- fluidPage(
    wired_search(
      inputId = "srch",
      label = "Text (Enter to update):",
      placeholder = "Write something!"
    ),
    verbatimTextOutput(outputId = "res")
  )

  server <- function(input, output, session) {

    output$res <- renderPrint(input$srch)

  }

  shinyApp(ui, server)

}
}
