% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{wired_slider}
\alias{wired_slider}
\title{Wired Slider}
\usage{
wired_slider(inputId, label = NULL, value = 10, min = 0, max = 100,
  step = 1, radius = 20, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{The initial value of the slider.}

\item{min}{The minimum value (inclusive) that can be selected.}

\item{max}{The maximum value (inclusive) that can be selected.}

\item{step}{Specifies the interval between each selectable value on the slider.}

\item{radius}{Radius of the knob of the slider.}

\item{width}{The width of the input, e.g. \code{400px}, or \code{100\%}; see \link[htmltools]{validateCssUnit}.}
}
\description{
Wired Slider
}
\examples{
if (interactive()) {

  library(shiny)
  library(wired)

  ui <- fluidPage(
    wired_slider(
      inputId = "sldr",
      label = "Slider:",
      min = 50,
      max = 600,
      value = 300,
      step = 10,
      radius = 10
    ),
    verbatimTextOutput(outputId = "res")
  )

  server <- function(input, output, session) {

    output$res <- renderPrint(input$sldr)

  }

  shinyApp(ui, server)

}
}
