% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{wired_text}
\alias{wired_text}
\title{Wired Text Input}
\usage{
wired_text(inputId, label = NULL, placeholder = NULL, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control.}

\item{width}{The width of the input, e.g. \code{400px}, or \code{100\%}; see \link[htmltools]{validateCssUnit}.}
}
\description{
Wired Text Input
}
\examples{
\dontrun{

if (interactive()) {



}

}
}
