% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{wired_tooltip}
\alias{wired_tooltip}
\title{Wired Tooltip}
\usage{
wired_tooltip(id, text, position = c("left", "right", "top", "bottom"))
}
\arguments{
\item{id}{Id of the element the tooltip is for.}

\item{text}{Text in the tooltip.}

\item{position}{Postion where tolltip will appear.}
}
\description{
Tooltip with text that appears on hover over an element.
}
\examples{
\dontrun{

if (interactive()) {

library(shiny)
library(wired)

ui <- fluidPage(
  fluidRow(
    column(
      width = 4,
      wired_slider(
        inputId = "n",
        label = "Number of obs. :",
        min = 50, value = 100, max = 300,
        radius = 10
      ),
      wired_tooltip(
        id = "n",
        text = "This is a slider !",
        position = "right"
      )
    ),
    column(
      width = 8,
      plotOutput("plot"),
      wired_tooltip(
        id = "plot",
        text = "And a plot !",
        position = "bottom"
      )
    )
  )
)

server <- function(input, output, session) {

  output$plot <- renderPlot({
    plot(sin(seq(from = pi, by = 0.05, length.out = input$n)))
  })

}

shinyApp(ui, server)

}

}
}
