% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{xkcd charts}
\usage{
xkcd(x, y, data = NULL, group = NULL, type = "line")
}
\arguments{
\item{x, y}{Values to use on x-axis / y-axis, can be a \code{character}
or a \code{formula} (\code{data} must be provided), or a \code{vector}.}

\item{data}{An object of type \code{data.frame}, or something than
be converted with \code{as.data.frame}.}

\item{group}{A grouping variable, same specification than \code{x} or \code{y}.}

\item{type}{Type of chart : line, scatter, bar, pie or radar.}
}
\value{
an \code{xkcd} htmlwidget object.
}
\description{
Htmlwidget for Chart.xkcd (\url{https://timqian.com/chart.xkcd})
 to create "sketchy", "cartoony" or "hand-drawn" styled charts.
}
\examples{

# With vectors
xkcd(x = 1:10, y = exp(1:10))

# With a data.frame and formulas
dat <- data.frame(
  x = c(1:10, 1:10),
  y = c(exp(1:10), (1:10)^4),
  group = rep(c("exp", "power4"), each = 10)
)

xkcd(x = ~x, y = ~y, group = ~group, data = dat)



# Scatter plot
xkcd(
  x = ~Sepal.Length,
  y = ~Sepal.Width,
  data = iris,
  type = "scatter"
)

xkcd(
  x = ~Sepal.Length,
  y = ~Sepal.Width,
  group = ~Species,
  data = iris,
  type = "scatter"
)


# Bar chart
dat <- data.frame(
  x = c("sketchy", "cartoony", "hand-drawn"),
  y = runif(3, 10, 90)
)
xkcd(x = ~x, y = ~y, data = dat, type = "bar")


# Pie chart
xkcd(x = ~x, y = ~y, data = dat, type = "pie")


# Radar chart
dat <- data.frame(
  x = rep(c("c", "h", "a", "r", "t"), times = 2),
  y = runif(10, 10, 90),
  group = rep(c("A", "B"), each = 5)
)
xkcd(x = ~x, y = ~y, group = ~group, data = dat, type = "radar")






}
