% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd_chart.R
\name{xkcd_chart}
\alias{xkcd_chart}
\title{<Add Title>}
\usage{
xkcd_chart(config, type = c("line", "scatter", "bar", "pie", "radar"),
  width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}
