% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd_labs.R
\name{xkcd_labs}
\alias{xkcd_labs}
\title{Add labs to a xkcd chart}
\usage{
xkcd_labs(xkcd, title = NULL, x = NULL, y = NULL)
}
\arguments{
\item{xkcd}{An \code{xkcd} htmlwidgets objet.}

\item{title}{The text for the title.}

\item{x}{The text for the x-axis title.}

\item{y}{The text for the y-axis title.}
}
\value{
an \code{\link{xkcd}} htmlwidget object.
}
\description{
Add labs to a xkcd chart
}
\examples{
xkcd(x = 1:10, y = exp(1:10)) \%>\%
  xkcd_labs(title = "The title of the chart",
            x = "X label",
            y = "Y label")
}
