% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd_options.R
\name{xkcd_options}
\alias{xkcd_options}
\title{Set options for an xkcd chart}
\usage{
xkcd_options(
  xkcd,
  dataColors = NULL,
  showLine = NULL,
  timeFormat = NULL,
  showLegend = NULL,
  dotSize = NULL,
  xTickCount = NULL,
  yTickCount = NULL,
  ticksCount = NULL,
  innerRadius = NULL,
  showLabels = NULL,
  fontFamily = NULL,
  unxkcdify = NULL,
  strokeColor = NULL,
  backgroundColor = NULL
)
}
\arguments{
\item{xkcd}{An \code{xkcd} htmlwidgets objet.}

\item{dataColors}{Array of colors for different datasets.}

\item{showLine}{(scatter) Connect the points with lines (default: false).}

\item{timeFormat}{(scatter) specify the time format if the x values are time (default undefined)
chart.xkcd use dayjs to format time, you can find the all the available formats
\href{https://github.com/iamkun/dayjs/blob/dev/docs/en/API-reference.md#list-of-all-available-formats}{here}.}

\item{showLegend}{(radar) Display legend near chart (default false)}

\item{dotSize}{(scatter, radar) You can change size of the dots if you want (default 1).}

\item{xTickCount}{(scatter) Customize tick numbers you want to see on the x axis (default 3)}

\item{yTickCount}{(scatter, line, bar) Customize tick numbers you want to see on the y axis (default 3)}

\item{ticksCount}{(radar) Customize tick numbers you want to see on the main line (default 3).}

\item{innerRadius}{(radar) Specify empty pie chart radius (default: 0.5), want a pie chart? set innerRadius to 0.}

\item{showLabels}{(radar) Display labels near every line (default false)}

\item{fontFamily}{Customize font family used in the chart}

\item{unxkcdify}{Disable xkcd effect (default false)}

\item{strokeColor}{Stroke colors (default black).}

\item{backgroundColor}{Color for BG (default white).}
}
\value{
an \code{\link{xkcd}} htmlwidget object.
}
\description{
Set options for an xkcd chart
}
\examples{

# change colors
xkcd(x = 1:10, y = exp(1:10)) \%>\%
  xkcd_options(dataColors = "blue")

# more classic
xkcd(x = 1:10, y = exp(1:10)) \%>\%
  xkcd_options(unxkcdify = TRUE)


# Timeserie
dat <- data.frame(
  time = Sys.Date() + 1:10,
  value = exp(1:10)
)
xkcd(x = ~time, y = ~value, data = dat, type = "scatter") \%>\%
  xkcd_options(timeFormat = "YYYY-MM-DD", showLine = TRUE)


# Pie chart
dat <- data.frame(
  x = c("sketchy", "cartoony", "hand-drawn"),
  y = runif(3, 10, 90)
)
xkcd(x = ~x, y = ~y, data = dat, type = "pie") \%>\%
  xkcd_options(innerRadius = 0)

}
