% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{wired_card}
\alias{wired_card}
\title{Wired Card}
\usage{
wired_card(..., elevation = 2, padding = 10, width = NULL)
}
\arguments{
\item{...}{UI objects to be contained in the card.}

\item{elevation}{Numerical number between 1-5 (inclusive) - sets the elevation of the card. Default is 2.}

\item{padding}{Amount of padding to apply. Can be numeric (in pixels) or character (e.g. "3em").}

\item{width}{The width of the input, e.g. \code{400px}, or \code{100\%}; see \link[htmltools]{validateCssUnit}.}
}
\description{
Wired Card
}
\examples{
\dontrun{

if (interactive()) {



}

}
}
