% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{wired_icon_button}
\alias{wired_icon_button}
\title{Wired Icon Button}
\usage{
wired_icon_button(inputId, icon, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value. The \code{input} slot that will be used to access the value.}

\item{icon}{An \code{\link[shiny:icon]{shiny::icon}} to appear on the button.}

\item{...}{Named attributes to be applied to the button.}
}
\description{
Wired Icon Button
}
\examples{
if (interactive()) {

  library(shiny)
  library(wired)

  ui <- fluidPage(
    wired_icon_button(
      inputId = "go",
      icon = icon("close")
    ),
    tags$br(),
    verbatimTextOutput(outputId = "res")
  )

  server <- function(input, output, session) {

    output$res <- renderPrint(input$go)

  }

  shinyApp(ui, server)

}
}
